pyproject.toml
[project]
name = "hydra-auto-schema"
description = "Automagically generate yaml schemas for your Hydra config files"
readme = "README.md"
authors = [{ name = "Fabrice Normandin", email = "normandf@mila.quebec" }]
requires-python = ">=3.10"
dependencies = [
    "docstring-parser>=0.16",
    "hydra-core>=1.3.2",
    "hydra-zen>=0.13.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.9.2",
    "rich>=13.9.3",
    "tqdm>=4.66.6",
    "watchdog>=5.0.3",
]
dynamic = ["version"]
[project.scripts]
hydra-auto-schema = "hydra_auto_schema.__main__:main"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"


[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.pytest.ini_options]
testpaths = ["hydra_auto_schema", "tests"]
norecursedirs = [".venv"]
addopts = ["--doctest-modules"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-regressions>=2.5.0",
    "uv-dynamic-versioning>=0.2.0",
    "pytest-cov>=5.0.0",
    "pre-commit<4.0.0",
]
managed = true

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"

[tool.hatch.build.targets.wheel]
packages = ["hydra_auto_schema", "hydra_plugins"]
